requirements.txt
flask==2.0.1
opencv-python==4.8.0.74
numpy==1.24.3
imutils==0.5.4
pyttsx3==2.90
pillow==10.0.0
flask-socketio==5.1.1 